# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CommitRequest": {
        "additionalProperties": false,
        "properties": {
          "conflict_acceptance": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Acceptance"
          },
          "digest": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Digest",
            "type": "string"
          },
          "selected": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Batch drafts only: indexes of the ticked items.",
            "title": "Selected"
          }
        },
        "required": [
          "digest"
        ],
        "title": "CommitRequest",
        "type": "object"
      },
      "CommitResult": {
        "additionalProperties": false,
        "properties": {
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "saved_ids": {
            "description": "Batch drafts: every saved event or task ID.",
            "items": {
              "type": "string"
            },
            "title": "Saved Ids",
            "type": "array"
          },
          "status": {
            "const": "committed",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "event_id",
          "status"
        ],
        "title": "CommitResult",
        "type": "object"
      },
...
      "ConversationCreated": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "conversation_id",
          "revision"
        ],
        "title": "ConversationCreated",
        "type": "object"
      },
      "DiscardResult": {
        "additionalProperties": false,
        "properties": {
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "status": {
            "const": "discarded",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "status"
        ],
        "title": "DiscardResult",
        "type": "object"
      },
      "Draft": {
        "additionalProperties": false,
        "properties": {
          "anchor_at": {
            "format": "date-time",
            "title": "Anchor At",
            "type": "string"
          },
          "basis_phrase": {
            "title": "Basis Phrase",
            "type": "string"
          },
          "committed_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed Event Id"
          },
          "confirmable": {
            "title": "Confirmable",
            "type": "boolean"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "expired": {
            "title": "Expired",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "fields": {
            "$ref": "#/components/schemas/DraftFields"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BatchItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Batch drafts only: what was read from the image.",
            "title": "Items"
          },
          "kind": {
            "description": "create makes a new event; the others act on `target`. `fields` is the result after the change. The task_* kinds act on a flexible task, which has a deadline instead of a slot.",
            "enum": [
              "create",
              "change",
              "cancel",
              "excuse",
              "task_create",
              "task_change",
              "task_cancel",
              "batch"
            ],
            "title": "Kind",
            "type": "string"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "status": {
            "enum": [
              "needs_clarification",
              "ready",
              "committed",
              "superseded",
              "discarded"
            ],
            "title": "Status",
            "type": "string"
          },
          "superseded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DraftTarget"
              },
              {
                "$ref": "#/components/schemas/TaskTarget"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "draft_id",
          "conversation_id",
          "status",
          "digest",
          "fields",
          "missing",
          "basis_phrase",
          "anchor_at",
          "expires_at",
          "superseded_by",
          "committed_event_id",
          "expired",
          "confirmable",
          "kind",
          "target"
        ],
        "title": "Draft",
        "type": "object"
      },
      "DraftFields": {
        "additionalProperties": false,
        "properties": {
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Flexible tasks only: when the user wants it done by.",
            "title": "Deadline"
          },
          "end_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "precision": {
            "anyOf": [
              {
                "enum": [
                  "date",
                  "instant"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "date when the user named a day, instant when a clock time was named.",
            "title": "Precision"
          },
          "recurrence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recurrence"
              },
              {
                "type": "null"
              }
            ]
          },
          "start_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "title",
          "timezone",
          "start_at",
          "end_at",
          "location",
          "recurrence"
        ],
        "title": "DraftFields",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "acceptance_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "CONFLICT_REVIEW_REQUIRED only: send back to save despite the overlap.",
            "title": "Acceptance Token"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "conflicts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConflictPair"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "CONFLICT_REVIEW_REQUIRED only: candidate slot vs saved occurrence.",
            "title": "Conflicts"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "superseded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "DRAFT_SUPERSEDED only.",
            "title": "Superseded By"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ImageInput": {
        "additionalProperties": false,
        "properties": {
          "data_base64": {
            "maxLength": 8000000,
            "minLength": 1,
            "title": "Data Base64",
            "type": "string"
          },
          "mime_type": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Mime Type",
            "type": "string"
          }
        },
        "required": [
          "mime_type",
          "data_base64"
        ],
        "title": "ImageInput",
        "type": "object"
      },
...
      "Message": {
        "additionalProperties": false,
        "properties": {
          "action_results": {
            "items": {
              "$ref": "#/components/schemas/ToolResult"
            },
            "title": "Action Results",
            "type": "array"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "message_id",
          "sequence",
          "role",
          "content",
          "created_at",
          "action_results",
          "draft_id"
        ],
        "title": "Message",
        "type": "object"
      },
      "MessagePage": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Draft"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "pending_client_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "An unfinished turn. Resend with this ID to resume it instead of starting another.",
            "title": "Pending Client Message Id"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "conversation_id",
          "revision",
          "items",
          "next_cursor",
          "pending_client_message_id",
          "draft"
        ],
        "title": "MessagePage",
        "type": "object"
      },
...
      "SendMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "client_message_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Client Message Id",
            "type": "string"
          },
          "content": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "image": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "client_message_id",
          "content",
          "timezone",
          "expected_revision"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
...
      "ToolResult": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "title": "Data"
          },
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "rejected",
              "clarify",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status",
          "data",
          "draft_id"
        ],
        "title": "ToolResult",
        "type": "object"
      },
      "TurnResult": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "assistant_message_id": {
            "title": "Assistant Message Id",
            "type": "string"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Draft"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "tool_results": {
            "items": {
              "$ref": "#/components/schemas/ToolResult"
            },
            "title": "Tool Results",
            "type": "array"
          },
          "user_message_id": {
            "title": "User Message Id",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "revision",
          "user_message_id",
          "assistant_message_id",
          "answer",
          "tool_results",
          "draft"
        ],
        "title": "TurnResult",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kairos API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/conversations": {
      "post": {
        "operationId": "start_api_v1_conversations_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/conversations/{conversation_id}/messages": {
      "get": {
        "operationId": "history_api_v1_conversations__conversation_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": -1,
              "minimum": -1,
              "title": "Cursor",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "History",
        "tags": [
          "assistant"
        ]
      },
      "post": {
        "operationId": "send_api_v1_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Send",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}": {
      "get": {
        "operationId": "draft_api_v1_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Draft",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/commit": {
      "post": {
        "operationId": "commit_api_v1_drafts__draft_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Commit",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/discard": {
      "post": {
        "operationId": "discard_api_v1_drafts__draft_id__discard_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscardResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Discard",
        "tags": [
          "assistant"
        ]
      }
    },
...
  }
}